+++ func/stats.js
@@ -27,6 +27,8 @@ function saveStats(level) {
 				stats.set("Manual-Unknown", stats.get("Manual-Unknown") + 1 || 1);
 			} else if (level == "revert") {
 				stats.set("Manual-Reversions", stats.get("Manual-Reversions") + 1 || 1);
+			} else if (level == "dm") {
+				stats.set("Direct-Message-Images", stats.get("Direct-Message-Images") + 1 || 1);
 			} else {
 				console.error(`[${dateToTime(new Date())}]: Error while saving the stats. Literally impossible to get to this, so if we have, something weird has happened.`);
 			}